refactor: 도메인 엔티티 패키지 구체화
