+++ src/main/java/roomescape/domain/reservation/slot/Theme.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation.slot;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
