+++ src/main/java/roomescape/domain/reservation/waiting/ReservationWaitingTicket.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation;
+package roomescape.domain.reservation.waiting;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -6,6 +6,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.OneToOne;
 import java.time.LocalDateTime;
+import roomescape.domain.reservation.Reservation;
 
 @Entity
 public class ReservationWaitingTicket {
