+++ src/main/java/roomescape/dto/time/AvailableReservationTimeResponseDto.java
@@ -2,7 +2,7 @@ package roomescape.dto.time;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.reservation.slot.ReservationTime;
 
 public record AvailableReservationTimeResponseDto(
         Long id,
