+++ src/main/java/roomescape/repository/JpaReservationTimeRepository.java
@@ -2,7 +2,7 @@ package roomescape.repository;
 
 import java.time.LocalTime;
 import org.springframework.data.jpa.repository.JpaRepository;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.reservation.slot.ReservationTime;
 
 public interface JpaReservationTimeRepository extends JpaRepository<ReservationTime, Long> {
     boolean existsByStartAt(LocalTime startAt);
